Renaming the application
