+++ android/app/BUCK
@@ -35,12 +35,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "edu.mit.privatekit",
+    package = "com.covidsafepaths",
 )
 
 android_resource(
     name = "res",
-    package = "edu.mit.privatekit",
+    package = "com.covidsafepaths",
     res = "src/main/res",
 )
