+++ android/app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">PrivateKit</string>
+    <string name="app_name">COVID Safe Paths</string>
 </resources>
