+++ android/settings.gradle
@@ -1,4 +1,4 @@
-rootProject.name = 'PrivateKit'
+rootProject.name = 'COVIDSafePaths'
 include ':react-native-svg'
 project(':react-native-svg').projectDir = new File(rootProject.projectDir, '../node_modules/react-native-svg/android')
 include ':react-native-fs'
