+++ app.json
@@ -1,4 +1,4 @@
 {
-  "name": "PrivateKit",
-  "displayName": "PrivateKit"
+  "name": "COVIDSafePaths",
+  "displayName": "COVID Safe Paths"
 }
