+++ ios/COVIDSafePaths.xcodeproj/xcshareddata/xcschemes/COVIDSafePaths.xcscheme
@@ -29,9 +29,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-               BuildableName = "PrivateKit.app"
-               BlueprintName = "PrivateKit"
-               ReferencedContainer = "container:PrivateKit.xcodeproj">
+               BuildableName = "COVIDSafePaths.app"
+               BlueprintName = "COVIDSafePaths"
+               ReferencedContainer = "container:COVIDSafePaths.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
          <BuildActionEntry
@@ -43,9 +43,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "00E356ED1AD99517003FC87E"
-               BuildableName = "PrivateKitTests.xctest"
-               BlueprintName = "PrivateKitTests"
-               ReferencedContainer = "container:PrivateKit.xcodeproj">
+               BuildableName = "COVIDSafePathsTests.xctest"
+               BlueprintName = "COVIDSafePathsTests"
+               ReferencedContainer = "container:COVIDSafePaths.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
       </BuildActionEntries>
@@ -59,9 +59,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "PrivateKit.app"
-            BlueprintName = "PrivateKit"
-            ReferencedContainer = "container:PrivateKit.xcodeproj">
+            BuildableName = "COVIDSafePaths.app"
+            BlueprintName = "COVIDSafePaths"
+            ReferencedContainer = "container:COVIDSafePaths.xcodeproj">
          </BuildableReference>
       </MacroExpansion>
       <Testables>
@@ -70,9 +70,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "00E356ED1AD99517003FC87E"
-               BuildableName = "PrivateKitTests.xctest"
-               BlueprintName = "PrivateKitTests"
-               ReferencedContainer = "container:PrivateKit.xcodeproj">
+               BuildableName = "COVIDSafePathsTests.xctest"
+               BlueprintName = "COVIDSafePathsTests"
+               ReferencedContainer = "container:COVIDSafePaths.xcodeproj">
             </BuildableReference>
          </TestableReference>
       </Testables>
@@ -92,9 +92,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "PrivateKit.app"
-            BlueprintName = "PrivateKit"
-            ReferencedContainer = "container:PrivateKit.xcodeproj">
+            BuildableName = "COVIDSafePaths.app"
+            BlueprintName = "COVIDSafePaths"
+            ReferencedContainer = "container:COVIDSafePaths.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </LaunchAction>
@@ -109,9 +109,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "PrivateKit.app"
-            BlueprintName = "PrivateKit"
-            ReferencedContainer = "container:PrivateKit.xcodeproj">
+            BuildableName = "COVIDSafePaths.app"
+            BlueprintName = "COVIDSafePaths"
+            ReferencedContainer = "container:COVIDSafePaths.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </ProfileAction>
