+++ ios/COVIDSafePaths.xcworkspace/xcshareddata/IDEWorkspaceChecks.plist
