+++ ios/COVIDSafePaths/AppDelegate.h
