+++ ios/COVIDSafePaths/Images.xcassets/AppIcon.appiconset/Contents.json
