+++ ios/COVIDSafePaths/Images.xcassets/AppIcon.appiconset/ShieldKeyHole1024x1024-20@3x.png
