+++ ios/COVIDSafePaths/Images.xcassets/AppIcon.appiconset/ShieldKeyHole1024x1024-40.png
