+++ ios/COVIDSafePaths/Images.xcassets/AppIcon.appiconset/ShieldKeyHole1024x1024-60@2x.png
