+++ ios/COVIDSafePaths/Images.xcassets/AppIcon.appiconset/ShieldKeyHole1024x1024-76.png
