+++ ios/COVIDSafePaths/Images.xcassets/Contents.json
