+++ ios/COVIDSafePathsTests/Info.plist
